refactor(schemas): split remaining domains into folders
