+++ packages/schemas/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./constants"
-export * from "./subject"
-export * from "./guide"
+export * from "./subjects"
+export * from "./guides"
 export * from "./identity"
-export * from "./learning-path"
+export * from "./learning-paths"
