+++ packages/schemas/src/learning-paths/index.ts
@@ -1,0 +1,3 @@
+export * from "./fields"
+export * from "./requests"
+export * from "./responses"
