+++ packages/schemas/src/subjects/fields.ts
@@ -1,0 +1,6 @@
+import { z } from "zod"
+
+export const subjectNameSchema = z
+  .string()
+  .min(3, { message: "Subject must be at least 3 characters long." })
+  .max(35, { message: "Subject can be no longer than 35 characters." })
